+++ ft_lstsize_bonus.c
@@ -6,7 +6,7 @@
 /*   By: vseppane <vseppane@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/25 15:30:40 by vseppane          #+#    #+#             */
-/*   Updated: 2024/04/25 15:30:52 by vseppane         ###   ########.fr       */
+/*   Updated: 2024/05/04 16:06:47 by vseppane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (lst)
