+++ ft_putendl_fd.c
@@ -6,7 +6,7 @@
 /*   By: vseppane <vseppane@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/21 14:55:42 by vseppane          #+#    #+#             */
-/*   Updated: 2024/04/24 11:57:59 by vseppane         ###   ########.fr       */
+/*   Updated: 2024/05/04 16:14:33 by vseppane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
