+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: vseppane <vseppane@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/21 15:29:44 by vseppane          #+#    #+#             */
-/*   Updated: 2024/04/25 15:23:58 by vseppane         ###   ########.fr       */
+/*   Updated: 2024/05/04 16:23:00 by vseppane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (n > 0)
